backend/requirements.txt
flask
flask-cors
joblib
scikit-learn
numpy
torch
transformers
python-dotenv
openai
